Se arreglo responsividad dash, tareas fallos de boton, notas validaciones

+++ src/components/AdminDash/UsersCrud.js
@@ -160,6 +160,16 @@ function UsersCrud() {
     };
 
     const handleDeleteUser = async (userId) => {
+        if(Number(localStorage.getItem("Usuario")) === userId){
+            Swal.fire({
+                icon:"warning",
+                title: 'No te puedes eliminar a ti mismo',
+                showDenyButton: false,
+                showCancelButton: false,
+                confirmButtonText: 'Ok',
+            })
+            return;
+        }
         try {
             await axios.delete(`http://localhost:4000/Usuarios/${userId}`); // Ajusta la URL según la configuración de tu servidor
             // Actualizar la lista de usuarios después de la eliminación exitosa
@@ -188,17 +198,17 @@ function UsersCrud() {
 
     return (
         <div className="container mx-auto p-4">
-            <h2 className="text-2xl font-bold mb-4">Tabla de Usuarios</h2>
+            <h2 className="md:text-5xl text-3xl font-bold text-center italic text-zinc-800">Gestión de usuarios</h2>
 
             {/* Formulario para crear usuarios */}
-            <Modal isOpen={modalIsOpen} onRequestClose={cerrarModal} >
-                <button className=' text-3xl font-bold' onClick={cerrarModal}> x </button>
-                <h1 className='text-2xl font-bold mb-4 text-center'> Crear Usuario </h1>
+            <Modal isOpen={modalIsOpen} onRequestClose={cerrarModal} className="w-8/12 m-auto my-10 rounded-xl  bg-zinc-50">
+                <button className=' text-3xl font-bold m-5' onClick={cerrarModal}> x </button>
+                <h1 className="md:text-5xl text-3xl font-bold text-center italic text-zinc-800">Agregar usuario</h1>
                 <form className='mt-10 flex flex-col' onSubmit={crearUsuarioForm}>
                     <input
                         type="text"
                         minLength={5}
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500" // Agrega las clases de Tailwind para estilos
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700"
                         placeholder="Nombre Completo"
                         required
                         value={usuarios.Nombre}
@@ -206,7 +216,7 @@ function UsersCrud() {
                     />
                     <input
                         type="email"
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500" // Agrega las clases de Tailwind para estilos
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700"
                         placeholder="Correo Electrónico"
                         value={usuarios.Correo}
                         onChange={(e) => setUsuarios({ ...usuarios, Correo: e.target.value })}
@@ -215,7 +225,7 @@ function UsersCrud() {
                     <input
                         type="password"
                         minLength={8}
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500" // Agrega las clases de Tailwind para estilos
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700"
                         placeholder="Contraseña"
                         value={usuarios.Contrasena}
                         onChange={(e) => setUsuarios({ ...usuarios, Contrasena: e.target.value })}
@@ -223,35 +233,35 @@ function UsersCrud() {
                     />
                     <input
                         type="password"
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500"
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700"
                         placeholder="Confirmar Contraseña"
                         value={usuarios.ConfirmarContrasena}
                         onChange={(e) => setUsuarios({ ...usuarios, ConfirmarContrasena: e.target.value })}
                         required
                     />
                     <select
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500"
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700"
                         onChange={(e) => setUsuarios({ ...usuarios, FK_Tipo_Usuario: e.target.value })}
                     >
-                        <option value="2">Administrador</option>
-                        <option value="1">Usuario</option>
+                        <option value="2" className='font-medium text-lg text-zinc-800 bg-gray-200'>Administrador</option>
+                        <option value="1" selected className='font-medium text-lg text-zinc-800 bg-gray-200'>Usuario</option>
                     </select>
-                    <div className='flex justify-center mt-5'>
-                        <button type='submit' className="bg-[#F95757] text-white py-2 w-40 rounded-xl font-bold"> Registrarse </button>
+                    <div className='flex justify-center'>
+                        <button type='submit' className="bg-[#1F4172] hover:bg-blue-600 text-white py-2 w-40 rounded-xl font-bold m-5"> Registrar usuario </button>
                     </div>
 
                 </form>
             </Modal>
 
             {/* Formulario para actualizar usuarios */}
-            <Modal isOpen={updateModalIsOpen} onRequestClose={cerrarUpdateModal}>
-            <button className=' text-3xl font-bold' onClick={cerrarUpdateModal}> x </button>
-                <h1 className='text-2xl font-bold mb-4 text-center'> Crear Usuario </h1>
+            <Modal isOpen={updateModalIsOpen} onRequestClose={cerrarUpdateModal} className="w-8/12 m-auto my-20 rounded-xl  bg-zinc-50">
+                <button className=' text-3xl font-bold mx-5' onClick={cerrarUpdateModal}> x </button>
+                <h1 className="md:text-5xl text-3xl font-bold text-center italic text-zinc-800">Editar usuario</h1>
                 <form className='mt-10 flex flex-col' onSubmit={actualizarUsuarioForm}>
                     <input
                         type="text"
                         minLength={5}
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500" // Agrega las clases de Tailwind para estilos
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700" // Agrega las clases de Tailwind para estilos
                         placeholder="Nombre Completo"
                         required
                         value={updateUsuario.Nombre}
@@ -259,7 +269,7 @@ function UsersCrud() {
                     />
                     <input
                         type="email"
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500" // Agrega las clases de Tailwind para estilos
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700" // Agrega las clases de Tailwind para estilos
                         placeholder="Correo Electrónico"
                         value={updateUsuario.Correo}
                         onChange={(e) => setUpdateUsuario({ ...updateUsuario, Correo: e.target.value })}
@@ -268,22 +278,22 @@ function UsersCrud() {
                     <input
                         type="password"
                         minLength={8}
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500" // Agrega las clases de Tailwind para estilos
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700" // Agrega las clases de Tailwind para estilos
                         placeholder="Contraseña"
                         value={updateUsuario.Contrasena}
                         onChange={(e) => setUpdateUsuario({ ...updateUsuario, Contrasena: e.target.value })}
                         required
                     />
                     <select
-                        className="w-auto p-2 mb-4 border text-black bg-gray-200 border-gray-300 rounded-3xl placeholder-gray-500"
+                        className="w-10/12 mx-auto p-2 mb-4 border font-medium text-lg text-zinc-800 bg-gray-200 border-gray-300 rounded-xl outline-none placeholder:text-zinc-700"
                         onChange={(e) => setUpdateUsuario({ ...updateUsuario, FK_Tipo_Usuario: e.target.value })}
                         value={updateUsuario.FK_Tipo_Usuario}
                     >
-                        <option value="2">Administrador</option>
-                        <option value="1">Usuario</option>
+                        <option value="2" className='font-medium text-lg text-zinc-800 bg-gray-200'>Administrador</option>
+                        <option value="1" className='font-medium text-lg text-zinc-800 bg-gray-200'>Usuario</option>
                     </select>
-                    <div className='flex justify-center mt-5'>
-                        <button type='submit' className="bg-[#427D9D] text-white py-2 w-40 rounded-xl font-bold"> Actualizar </button>
+                    <div className='flex justify-center align-middle items-center'>
+                        <button type='submit' className="bg-[#1F4172] hover:bg-blue-600 text-white py-2 w-40  rounded-xl font-bold m-5"> Actualizar </button>
                     </div>
 
                 </form>
@@ -292,41 +302,44 @@ function UsersCrud() {
 
             {/* Lista de usuarios */}
             <div className="container mx-auto p-4">
-                <div className='flex justify-between mx-6 mb-3'>
-                    <h3 className="text-lg font-semibold mb-2">Lista de Usuarios</h3>
-                    <button className='p-2 bg-blue-400 rounded-lg text-white' onClick={abrirModal}> añadir usuario</button>
+                <div className='flex justify-between mx-6 my-5'>
+                    <h3 className="md:text-2xl text-xl font-semibold italic my-auto">Lista de Usuarios</h3>
+                    <button className='bg-[#1F4172] font-medium text-lg text-white py-2 px-4 w-70 rounded-xl hover:bg-blue-600' onClick={abrirModal}> añadir usuario</button>
                 </div>
 
 
                 {users.length > 0 ?
                     (<>
-                        <table className="min-w-full divide-y divide-gray-300">
-                            <thead>
-                                <tr>
-                                    <th className="py-2 px-4 text-left bg-gray-200">Nombre</th>
-                                    <th className="py-2 px-4 text-left bg-gray-200">Correo</th>
-                                    <th className="py-2 px-4 text-left bg-gray-200">Acciones</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                {users.map((user, index) => (
-                                    <tr key={user.Id_Usuario} className={index % 2 === 0 ? 'bg-gray-100' : 'bg-white'}>
-                                        <td className="py-2 px-4">{user.Nombre}</td>
-                                        <td className="py-2 px-4">{user.Correo}</td>
-                                        <td className="py-2 px-4">
-                                            {user.FK_Tipo_Usuario === 1 ? (
-                                                <button className="bg-red-500 text-white px-2 py-1 rounded cursor-pointer" onClick={() => handleDeleteUser(user.Id_Usuario)}>Eliminar</button>
-                                            ) : null}
-                                            <button className="bg-blue-500 text-white px-2 py-1 rounded cursor-pointer ml-2" onClick={() => abrirUpdateModal(user)}>Actualizar</button>
-                                        </td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                        </table>
+<div className="overflow-x-auto">
+  <table className="min-w-full divide-y divide-gray-300">
+    <thead>
+      <tr>
+        <th className="py-2 px-4 text-center font-semibold italic my-auto bg-gray-200">Nombre</th>
+        <th className="py-2 px-4 text-center font-semibold italic my-auto bg-gray-200">Correo</th>
+        <th className="py-2 px-4 text-center font-semibold italic my-auto bg-gray-200">Acciones</th>
+      </tr>
+    </thead>
+    <tbody>
+      {users.map((user, index) => (
+        <tr key={user.Id_Usuario} className={index % 2 === 0 ? 'bg-gray-100' : 'bg-white'}>
+          <td className="py-2 text-center font-normal italic my-auto px-4 sm:px-2 md:px-4">{user.Nombre}</td>
+          <td className="py-2 text-center font-normal italic my-auto px-4 sm:px-2 md:px-4">{user.Correo}</td>
+          <td className="py-2 text-center font-normal italic my-auto px-4 sm:px-2 md:px-4">
+            {user.FK_Tipo_Usuario === 1 ? (
+              <button className="bg-[#B31312] text-white px-2 py-1 rounded cursor-pointer font-semibold md:my-auto my-2 mx-3" onClick={() => handleDeleteUser(user.Id_Usuario)}>Eliminar</button>
+            ) : null}
+            <button className="bg-[#1F4172] text-white px-2 py-1 rounded cursor-pointer md:my-auto font-semibold my-2" onClick={() => abrirUpdateModal(user)}>Editar</button>
+          </td>
+        </tr>
+      ))}
+    </tbody>
+  </table>
+</div>
+
 
                     </>) :
                     (<>
-                        <p className=' text-6xl text-center font-bold'> Todavia no hay usuarios añadidos </p>
+                        <p className='md:text-2xl text-xl font-semibold italic my-auto text-center'> Todavia no hay usuarios añadidos </p>
                     </>)}
